Broadcast chunk changes from non-entity ticking chunks
One of the behaviors of the current vanilla simulation
distance is that all chunks with changes are to be broadcasted,
regardless of their ticking level. To keep this behavior, as well
as the performance of the old behavior, a separate set of
chunks requiring broadcasting is maintained.

Currently, iterating over the entire set requires moving
the main tickChunks loop off of entity ticking and onto
just sending, but that would require inserting ticking
level checks - so while iteration cost would remain
roughly a little worse, the ticking checks might
cause trouble. So, keeping the main loop on entity
but maintaining a separate loop for just changes
will work without additional performance loss,
as the size of the broadcast set should be very
small.